==> aou/4_monitor_cloud_analysis_environment.ipynb <==
**Authors:** Nicole Deflaux

**Contributors:** Jennifer Zhang

# Monitor the Cloud Analysis Environment

<div class="alert alert-block alert-success">
It is helpful to view CPU utilization and memory utilization when your Cloud analysis environment is running. For example, if you are using a Dataproc cluster and you see that the CPU utilization of the worker machines is close to 100%, you could increase the number of workers to speed up the "wall clock time" it takes to complete the analysis.
</div>

**Usage**: 

1. Copy this notebook to the workspcae you wish to check the computing resources unilization. Then run the notebook and click on the links from the output to check the utilization plots.

1. It only works for when you are the of the workspace.

1. If you are only a WRITER, and not an OWNER, of the workspace in which your Cloud analysis environment is running, these utilization plots will say "You don't have access to this data."

import os

CPU_UTILIZATION = f'https://console.cloud.google.com/monitoring/metrics-explorer?authuser=0&project={os.getenv("GOOGLE_PROJECT")}&pageState=%7B%22xyChart%22:%7B%22dataSets%22:%5B%7B%22timeSeriesFilter%22:%7B%22filter%22:%22metric.type%3D%5C%22compute.googleapis.com%2Finstance%2Fcpu%2Futilization%5C%22%20resource.type%3D%5C%22gce_instance%5C%22%22,%22minAlignmentPeriod%22:%2260s%22,%22aggregations%22:%5B%7B%22perSeriesAligner%22:%22ALIGN_MEAN%22,%22crossSeriesReducer%22:%22REDUCE_NONE%22,%22alignmentPeriod%22:%2260s%22,%22groupByFields%22:%5B%5D%7D,%7B%22crossSeriesReducer%22:%22REDUCE_NONE%22,%22alignmentPeriod%22:%2260s%22,%22groupByFields%22:%5B%5D%7D%5D%7D,%22targetAxis%22:%22Y1%22,%22plotType%22:%22LINE%22%7D%5D,%22options%22:%7B%22mode%22:%22COLOR%22%7D,%22constantLines%22:%5B%5D,%22timeshiftDuration%22:%220s%22,%22y1Axis%22:%7B%22label%22:%22y1Axis%22,%22scale%22:%22LINEAR%22%7D%7D,%22isAutoRefresh%22:true,%22timeSelection%22:%7B%22timeRange%22:%221h%22%7D%7D'
print(f'To see the CPU utilization of your Cloud analysis environment, click on this link {CPU_UTILIZATION}')

MEMORY_UTILIZATION = f'https://console.cloud.google.com/monitoring/metrics-explorer?authuser=0&project={os.getenv("GOOGLE_PROJECT")}&pageState=%7B%22xyChart%22:%7B%22dataSets%22:%5B%7B%22timeSeriesFilter%22:%7B%22filter%22:%22metric.type%3D%5C%22agent.googleapis.com%2Fmemory%2Fpercent_used%5C%22%20resource.type%3D%5C%22gce_instance%5C%22%20metric.label.%5C%22state%5C%22%3D%5C%22used%5C%22%22,%22minAlignmentPeriod%22:%2260s%22,%22aggregations%22:%5B%7B%22perSeriesAligner%22:%22ALIGN_MEAN%22,%22crossSeriesReducer%22:%22REDUCE_NONE%22,%22alignmentPeriod%22:%2260s%22,%22groupByFields%22:%5B%5D%7D,%7B%22crossSeriesReducer%22:%22REDUCE_NONE%22,%22alignmentPeriod%22:%2260s%22,%22groupByFields%22:%5B%5D%7D%5D%7D,%22targetAxis%22:%22Y1%22,%22plotType%22:%22LINE%22%7D%5D,%22options%22:%7B%22mode%22:%22COLOR%22%7D,%22constantLines%22:%5B%5D,%22timeshiftDuration%22:%220s%22,%22y1Axis%22:%7B%22label%22:%22y1Axis%22,%22scale%22:%22LINEAR%22%7D%7D,%22isAutoRefresh%22:true,%22timeSelection%22:%7B%22timeRange%22:%221h%22%7D%7D'
print(f'To see the memory utilization of your Cloud analysis environment, click on this link {MEMORY_UTILIZATION}')

MACHINE_COUNT = f'https://console.cloud.google.com/monitoring/metrics-explorer?authuser=0&project={os.getenv("GOOGLE_PROJECT")}&pageState=%7B%22xyChart%22:%7B%22dataSets%22:%5B%7B%22timeSeriesFilter%22:%7B%22filter%22:%22metric.type%3D%5C%22compute.googleapis.com%2Finstance%2Fuptime%5C%22%20resource.type%3D%5C%22gce_instance%5C%22%22,%22minAlignmentPeriod%22:%22300s%22,%22aggregations%22:%5B%7B%22perSeriesAligner%22:%22ALIGN_RATE%22,%22crossSeriesReducer%22:%22REDUCE_SUM%22,%22alignmentPeriod%22:%22300s%22,%22groupByFields%22:%5B%22resource.label.%5C%22project_id%5C%22%22%5D%7D,%7B%22crossSeriesReducer%22:%22REDUCE_NONE%22,%22alignmentPeriod%22:%2260s%22,%22groupByFields%22:%5B%5D%7D%5D%7D,%22targetAxis%22:%22Y1%22,%22plotType%22:%22LINE%22%7D%5D,%22options%22:%7B%22mode%22:%22COLOR%22%7D,%22constantLines%22:%5B%5D,%22timeshiftDuration%22:%220s%22,%22y1Axis%22:%7B%22label%22:%22y1Axis%22,%22scale%22:%22LINEAR%22%7D%7D,%22isAutoRefresh%22:true,%22timeSelection%22:%7B%22timeRange%22:%221d%22%7D%7D'
print(f'To see the count of machines in your Cloud analysis environment, click on this link {MACHINE_COUNT}')
